Added a macro to simplify <Arch>IntrinsicType definitions
